requirements.txt
PySide6>=6.6.0
openai>=1.0.0
groq>=0.4.0
gpt4all>=2.0.0
google-generativeai>=0.3.0
